requirements.txt
streamlit
openai
pandas
python-dotenv
langchain
langchain-openai
tiktoken
numpy